+++ src/dwarfspec/config_schema.lua
@@ -26,6 +26,7 @@ local RESERVED_COMMANDS = {
     protocol_version=true,
     redraw=true,
     root=true,
+    setGamePaused=true,
     setViewPos=true,
     stage_overlay_registration=true,
     type=true,